feat: Refactor to ResourceFilteringStore service
