--- addons/api/app/services/filterable-store.js
@@ -1,1 +1,0 @@
-export { default } from 'api/services/filterable-store';
