+++ addons/api/app/services/resource-filtering-store.js
@@ -1,0 +1,1 @@
+export { default } from 'api/services/resource-filtering-store';
